fix: 大版本安全加固 v1.3.0 — 主机密钥验证/密钥库存储/随机盐导出/SFTP 独立会话/终端监听器泄漏/DB 事务与 WAL/CSP/日志系统 + 前端 20 余项泄漏与类型修复

+++ src-tauri/src/common.rs
@@ -31,7 +31,7 @@ impl std::fmt::Display for SessionId {
 pub struct OutputChunk {
     pub seq: u64,
     pub output: String,
-    pub ts: u128,
+    pub ts: u64,
 }
 
 impl OutputChunk {
@@ -39,7 +39,7 @@ impl OutputChunk {
         let ts = std::time::SystemTime::now()
             .duration_since(std::time::UNIX_EPOCH)
             .unwrap_or_default()
-            .as_millis();
+            .as_millis() as u64;
         Self { seq, output, ts }
     }
 }
